requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
websockets>=12.0
MetaTrader5>=5.0.45
pydantic>=2.0.0
